limit fixd size strings
