+++ Cargo.toml
@@ -43,6 +43,7 @@ thiserror = "1.0"
 encoding_rs = "0.8"
 uuid = "1.7"
 winstructs = "0.3"
+substring = "1.4"
 
 serde = { version = "1.0", features = ["derive"], optional=true }
 serde_json = {version = "1", optional=true}
